requirements-phase1.txt
openwakeword>=0.6.0
mlx-whisper>=0.4.0
psutil>=5.9.0
pyyaml>=6.0
sounddevice>=0.4.6
numpy>=1.24.0,<2.0.0
rich>=13.0.0
requests>=2.31.0
